+++ src/main/java/com/task_4_1/RequestIdComparator.java
@@ -7,6 +7,6 @@ import java.util.Comparator;
 public class RequestIdComparator implements Comparator<Request> {
     @Override
     public int compare(Request request1, Request request2) {
-        return request1.getId() - request2.getId();
+        return (int)(request1.getId() - request2.getId());
     }
 }
